+++ NEWS.md
@@ -9,6 +9,11 @@ editor_options:
 Add support for StellarNet, Brewer, and Bentham spectrometers (examples
 of data files are welcome for these and other instruments).
 
+# photobiologyInOut 0.4.26
+
+- Add function `read_spectrapen_csv()` supporting PSI's SpectraPen spectrometer.
+- Update for code-breaking changes in package 'readr'.
+
 # photobiologyInOut 0.4.25
 
 - Fix bug in `mspct2colorSpec()` which affected also